+++ src/test/java/codesquad/service/QnaServiceTest.java
@@ -3,6 +3,7 @@ package codesquad.service;
 import codesquad.CannotDeleteException;
 import codesquad.domain.Answer;
 import codesquad.domain.AnswerRepository;
+import codesquad.domain.Answers;
 import codesquad.domain.Question;
 import codesquad.domain.QuestionRepository;
 import codesquad.domain.User;
